fix test for use db, add __eq__ method to Creature model

+++ src/data/explorer.py
@@ -1,6 +1,7 @@
 from typing import Union, Optional
-from .init import curs
+from .init import conn, curs, IntegrityError
 from model.explorer import Explorer
+from error import Missing, Duplicate
 
 curs.execute(
     """create table if not exists explorer(
@@ -26,7 +27,7 @@ def get_one(name: str) -> Optional[Explorer]:
     curs.execute(qry, params)
     row = curs.fetchone()
     if row is None:
-        return None
+        raise Missing(msg=f"Explorer {name} not found")
     return row_to_model(row)
 
 
@@ -37,14 +38,21 @@ def get_all() -> list[Explorer]:
 
 
 def create(explorer: Explorer) -> Explorer:
+    if not explorer:
+        return None
     qry = """insert into explorer (name, country, description)
     values (:name, :country, :description)"""
     params = model_to_dict(explorer)
-    _ = curs.execute(qry, params)
+    try:
+        curs.execute(qry, params)
+    except IntegrityError:
+        raise Duplicate(msg=f"Explorer {explorer.name} already exists")
     return get_one(explorer.name)
 
 
 def modify(name: str, explorer: Explorer) -> Explorer:
+    if not (name and explorer):
+        return None
     qry = """update explorer
     set country=:country,
     name=:name,
@@ -52,13 +60,19 @@ def modify(name: str, explorer: Explorer) -> Explorer:
     where name=:name_orig"""
     params = model_to_dict(explorer)
     params["name_orig"] = explorer.name
-    _ = curs.execute(qry, params)
-    explorer2 = get_one(explorer.name)
-    return explorer2
+    curs.execute(qry, params)
+    if curs.rowcount == 1:
+        return get_one(explorer.name)
+    else:
+        raise Missing(msg=f"Explorer {name} not found")
 
 
-def delete(explorer: Explorer) -> bool:
+def delete(name: str) -> bool:
+    if not name:
+        return False
     qry = "delete from explorer where name = :name"
-    params = {"name": explorer.name}
+    params = {"name": name}
     res = curs.execute(qry, params)
+    if curs.rowcount != 1:
+        raise Missing(msg=f"Explorer {name} not found")
     return bool(res)
